+++ .eslintrc.js
@@ -1,7 +1,12 @@
 module.exports = {
-  extends: ['wolox-react'],
+  extends: ['wolox-react-native'],
   rules: {
+    complexity: 'off',
     'import/no-unresolved': 'off',
-    'import/no-extraneous-dependencies': 'off'
+    'import/no-extraneous-dependencies': 'off',
+    'no-magic-numbers': 'off',
+    'import/order': ['error', { 'newlines-between': 'always' }],
+    'no-nested-ternary': 'off',
+    'no-console': 'off'
   }
 };