+++ generators/app/index.js
@@ -8,7 +8,7 @@ const gitInitialization = require('./tasks/gitInitialization');
 const bitriseInitialization = require('./tasks/bitriseInitialization');
 const nextSteps = require('./tasks/nextSteps');
 const bundleInstall = require('./tasks/bundleInstall');
-const configurePods = require('./tasks/configurePods');
+const configureFastlane = require('./tasks/configureFastlane');
 const installPods = require('./tasks/installPods');
 const linkAppAssets = require('./tasks/linkAppAssets');
 
@@ -45,7 +45,10 @@ class ReactNativeBootstrap extends Generator {
           'Login',
           'Tabs',
           'Drawer',
+          'Redux Persist',
+          'Firebase Analytics',
           'Crashlytics',
+          'Firebase Performance',
           'Push Notifications',
           'Google Analytics',
           'Redux Persist',
@@ -87,7 +90,7 @@ class ReactNativeBootstrap extends Generator {
       .then(() => reactNativeCliInstall.bind(this)())
       .then(() => reactNativeInit.bind(this)())
       .then(() => installDependencies.bind(this)())
-      .then(() => configurePods.bind(this)());
+      .then(() => configureFastlane.bind(this)());
   }
 
   writing() {